+++ verticapy/learn/model_selection/__init__.py
@@ -14,4 +14,4 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.machine_learning.model_selection.model_selection import *
+from verticapy.machine_learning.model_selection import *
